Add files via upload
To run Saphyra you will need to select every "urllib2" repository in your Python Interpreter (in PyCharm you can access it by pressing ALT+CTRL+S), if some of them won't install it isn't a problem at all. You then will need to also select "sklearn" as another interpreter. 
You MUST have Python 3.6 selected, because it works only with this version and with 3.7. If you don't have it already, just download it here https://www.python.org/downloads/release/python-360/ . At the bottom of the pag select "Windows x86 executable installer" if your PC runs on Intel, "Windows x86-64 executable installer" if it uses AMD. 

Then, go to Python Interpreter and change on the first line Python by pressing +, and select Python 3.6 (if it is red, just press it so it can install it and after that select it again to use it).

And that's it, you're pretty much done, now you can DDoS a server by changing on the 18th line the URL. 

If you have ANY questions, i would be really happy to answer them on my instagram (avagliano_nicolo). I will reply as soon as possible.



This is PURELY for educational purpose, I do not take any responsabilities for what you use this code for. I just wanted to say this is NOT my code, its author is  "xdferlesh1", and i don't want to take any credit for what is not mine, but I thought that posting the changes I have done to the code to make it run on my computer.

Drop a follow both here and (if you need help) on Instagram.